requirements.txt
psutil==7.0.0
pydantic==2.11.3
pycryptodome==3.22.0
pyotp==2.9.0
qrcode==8.1
PyQt6==6.9.0
qt6-applications==6.5.0.2.3
pillow==11.2.1
pyinstaller==6.13.0
mypy==1.15.0
